requirements.txt
scikit-learn
pandas
matplotlib
seaborn
python-dotenv
gradio
google-generativeai
together
pyngrok
kagglehub
markdown
python-multipart
xhtml2pdf
joblib
requests
zipfile
numpy
ipykernel
Jinja2
tqdm 